# docs/swagger.yaml
basePath: /v2
definitions:
  models.ColumnsStruct:
    properties:
      columnName:
        type: string
      dataType:
        type: string
      isNullable:
        type: string
      primaryKey:
        type: boolean
    type: object
host: localhost:8080
info:
  contact:
    name: Dmitry
    url: https://github.com/dmitry-msk777/Connector_1C_Enterprise
  description: API description
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Swagger API Connector for 1C Enterprise
  version: "1.0"
paths:
  /getcounttable/{TableName}:
    get:
      consumes:
      - application/json
      description: Получить количество записей в таблице
      operationId: get-count-table
      parameters:
      - description: TableName
        in: path
        name: TableName
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          headers:
            TokenBearer:
              description: qwerty
              type: string
          schema:
            $ref: '#/definitions/models.ColumnsStruct'
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
        default:
          description: ""
          schema:
            type: string
      summary: GetCountTable
  /getdescription/{TableName}:
    get:
      consumes:
      - application/json
      description: Получить описание полей таблицы
      operationId: get-description
      parameters:
      - description: TableName
        in: path
        name: TableName
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          headers:
            TokenBearer:
              description: qwerty
              type: string
          schema:
            $ref: '#/definitions/models.ColumnsStruct'
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
        default:
          description: ""
          schema:
            type: string
      summary: GetDescription
swagger: "2.0"